[LG-416] Setup Forms for 2FA methods
**Why**:
We want to regularize the interfaces for our 2FA methods so it's
easier to add new 2FA methods.

**How**:
For this PR, we're creating a set of setup forms that present a
similar interface across methods.

[X] piv/cac
[X] totp
[ ] sms
[ ] voice
[ ] personal key
