+++ app/forms/two_factor_authentication/setup_form.rb
@@ -1,0 +1,16 @@
+module TwoFactorAuthentication
+  class SetupForm
+    include ActiveModel::Model
+
+    attr_accessor :user, :configuration_manager
+
+    validates :user, presence: true
+    validates :configuration_manager, presence: true
+
+    def method
+      @method ||= begin
+        self.class.name.demodulize.sub(/SetupForm$/, '').snakecase.to_sym
+      end
+    end
+  end
+end
